+++ plugins/commands/login/middleware/add_authentication.rb
@@ -18,7 +18,17 @@ module VagrantPlugins
 
           env[:box_urls].map! do |url|
             u = URI.parse(url)
-            if u.host == server_uri.host
+            replace = u.host == server_uri.host
+            if !replace
+              # We need this in here for the transition we made from
+              # Vagrant Cloud to Atlas. This preserves access tokens
+              # appending to both without leaking access tokens to
+              # unsavory URLs.
+              replace = u.host == "vagrantcloud.com" &&
+                server_uri.host == "atlas.hashicorp.com"
+            end
+
+            if replace
               u.query ||= ""
               u.query += "&" if u.query != ""
               u.query += "access_token=#{token}"